+++ Gmulator/Core.Gbc/Mappers/Mapper0.cs
@@ -6,16 +6,16 @@ public class Mapper0(byte[] rom, GbcMmu mmu) : BaseMapper(rom, mmu)
 
     public override void Init(byte[] rom, string filename) => base.Init(rom, filename);
 
-    public override int ReadRom(int a) => Rom[a % 0x4000];
+    public override byte ReadRom(int a) => Rom[a % 0x4000];
 
     public override Span<byte> ReadRomBlock(int a, int s) => new();
 
-    public override void WriteRom0(int a, int v)
+    public override void WriteRom0(int a, byte v)
     {
         //Rom[a] = v;
     }
 
-    public override void WriteRom1(int a, int v)
+    public override void WriteRom1(int a, byte v)
     {
         //if (edit)
         ///    Rom[a] = v;
