+++ Gmulator/Core.Gbc/Mappers/Mapper2.cs
@@ -10,7 +10,7 @@ public class Mapper2 : BaseMapper
 
     public override void Init(byte[] rom, string filename) => base.Init(rom, filename);
 
-    public override int ReadRom(int a)
+    public override byte ReadRom(int a)
     {
         var addr = a + (0x4000 * (Rombank - 1));
         if (Rombank > 1)
@@ -27,12 +27,12 @@ public class Mapper2 : BaseMapper
             return new(Rom, a + 0x4000 * (Rombank - 1), size);
     }
 
-    public override void WriteRom0(int a, int v)
+    public override void WriteRom0(int a, byte v)
     {
 
     }
 
-    public override void WriteRom1(int a, int v)
+    public override void WriteRom1(int a, byte v)
     {
         //if (edit)
         //    Rom[a + (0x4000 * (Rombank - 1))] = v;
