+++ Gmulator/Core.Nes/Sound/Dmc.cs
@@ -26,7 +26,7 @@ public class Dmc : BaseChannel
         nes.CpuMap.Set(0x00, 0x00, 0x4010, 0x4013, a => 0xff, Write, RamType.Register, 1);
     }
 
-    public void Write(int a, int v)
+    public void Write(int a, byte v)
     {
         switch (a)
         {
